Move all of the avatar-related functions to UserUtils
